shop/serializers.py
from rest_framework import serializers

from shop.models import Order


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = ('id', 'placed_at', 'status', 'customer')
