requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
aniso8601==9.0.1
    # via graphene
asgiref==3.7.2
    # via
    #   channels
    #   daphne
    #   django
attrs==23.1.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==23.6.2
    # via daphne
automat==22.10.0
    # via twisted
billiard==4.1.0
    # via celery
celery==5.3.4
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
channels[daphne]==4.0.0
    # via -r requirements.in
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
constantly==15.1.0
    # via twisted
coverage==7.3.1
    # via -r requirements.in
cryptography==41.0.4
    # via
    #   autobahn
    #   jwcrypto
    #   pyopenssl
    #   service-identity
daphne==4.0.0
    # via channels
deprecated==1.2.14
    # via jwcrypto
django==4.2.4
    # via
    #   -r requirements.in
    #   channels
    #   django-appconf
    #   django-ckeditor
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   django-guardian
    #   django-js-asset
    #   django-nine
    #   django-oauth-toolkit
    #   django-reversion
    #   djangorestframework
    #   drf-yasg
    #   easy-thumbnails
    #   graphene-django
django-appconf==1.0.5
    # via django-compressor
django-ckeditor==6.7.0
    # via -r requirements.in
django-compressor==4.4
    # via -r requirements.in
django-cors-headers==4.2.0
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-elasticsearch-dsl==7.3
    # via
    #   -r requirements.in
    #   django-elasticsearch-dsl-drf
django-elasticsearch-dsl-drf==0.22.5
    # via -r requirements.in
django-filter==23.3
    # via -r requirements.in
django-guardian==2.4.0
    # via -r requirements.in
django-js-asset==2.1.0
    # via django-ckeditor
django-nine==0.2.7
    # via django-elasticsearch-dsl-drf
django-oauth-toolkit==2.3.0
    # via -r requirements.in
django-reversion==5.0.6
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   django-elasticsearch-dsl-drf
    #   drf-yasg
drf-yasg==1.21.7
    # via -r requirements.in
easy-thumbnails==2.8.5
    # via -r requirements.in
elasticsearch==7.9.1
    # via
    #   django-elasticsearch-dsl-drf
    #   elasticsearch-dsl
elasticsearch-dsl==7.4.1
    # via
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
factory-boy==2.0.0
    # via -r requirements.in
faker==19.6.2
    # via -r requirements.in
graphene==3.3
    # via graphene-django
graphene-django==3.1.5
    # via -r requirements.in
graphql-core==3.2.3
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.2.0
    # via
    #   graphene
    #   graphene-django
gunicorn==21.2.0
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via twisted
inflection==0.5.1
    # via drf-yasg
iniconfig==2.0.0
    # via pytest
jwcrypto==1.5.0
    # via django-oauth-toolkit
kombu==5.3.2
    # via celery
oauthlib==3.2.2
    # via django-oauth-toolkit
packaging==23.1
    # via
    #   django-nine
    #   drf-yasg
    #   gunicorn
    #   pytest
pillow==10.0.0
    # via
    #   -r requirements.in
    #   easy-thumbnails
pluggy==1.3.0
    # via pytest
promise==2.3
    # via graphene-django
prompt-toolkit==3.0.39
    # via click-repl
psycopg==3.1.12
    # via -r requirements.in
psycopg2==2.9.7
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pyopenssl==23.2.0
    # via twisted
pytest==7.4.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   celery
    #   elasticsearch-dsl
    #   faker
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via drf-yasg
rcssmin==1.1.1
    # via django-compressor
redis==5.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
rjsmin==1.2.1
    # via django-compressor
sentry-sdk==1.30.0
    # via -r requirements.in
service-identity==23.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
    #   elasticsearch-dsl
    #   promise
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via graphene-django
twisted[tls]==23.8.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.8.0
    # via
    #   psycopg
    #   twisted
tzdata==2023.3
    # via celery
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.0.4
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.7
    # via prompt-toolkit
wrapt==1.15.0
    # via deprecated
zope-interface==6.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
